+++ tests/functional/ExampleTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace tests;
+
+/**
+ * Class ExampleTest
+ * @package tests
+ *
+ * @author Vasile Crudu <bazillio07@yandex.ru>
+ *
+ * @link https://github.com/vova07
+ */
+class ExampleTest extends TestCase
+{
+    /**
+     * Test function.
+     */
+    public function testExample()
+    {
+        $this->assertTrue(true);
+    }
+}